changed db to MYsql and changed the login page ui

+++ Emplyee_app/models.py
@@ -16,15 +16,14 @@ class Role(models.Model):
     def __str__(self):
         return self.name
 
-# Create your models here.
 class Employee(models.Model):
     first_name = models.CharField(max_length=100, null=False)
     last_name = models.CharField(max_length =100)
     dept = models.ForeignKey(Department, on_delete=models.CASCADE)
-    salary = models.IntegerField(default=0)
-    bonus = models.IntegerField(default=0)
+    salary = models.BigIntegerField(default=0)
+    bonus = models.BigIntegerField(default=0)
     role = models.ForeignKey(Role,on_delete=models.CASCADE)
-    phone = models.IntegerField(default=0)
+    phone = models.BigIntegerField(default=0)
     hire_date = models.DateField()
 
     def __str__(self):
